fix: move internal packages to pkg for Vercel compatibility
- Move all packages from internal/ to pkg/ directory to resolve Vercel build errors
- Update all import paths from neomovies-api/internal to neomovies-api/pkg
- Remove unused 'os' import from api/index.go
- Fix Go module visibility issues for serverless deployment
